third_party/WebKit/LayoutTests/fullscreen/full-screen-frameset-expected.txt

CONSOLE ERROR: line 28: Feature policy violation: fullscreen is not allowed in this document.


--------
Frame: 'header'
--------
Test for bug 58638: Do not allow fullscreen from within a <frame> element, even with the allowfullscreen attribute set.

To test manually, click the "Go full screen" button - the page should not enter full screen mode.

SUCCEED - did not enter full screen!
END OF TEST


--------
Frame: 'frame'
--------
Go full screen
